update docs, add kw constructor and JLSO test
