+++ test/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 IterationControl = "b3c1a2ee-3fec-4384-bf48-272ea71de57c"
+JLSO = "9da8a3cd-07a3-59c0-a743-3fdc52c30d11"
 MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
 MLJModelInterface = "e80e1ace-859a-464e-9ed9-23947d8ae3ea"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
