+++ test/runtests.jl
@@ -26,7 +26,3 @@ end
 @testset "traits" begin
     include("traits.jl")
 end
-
-@testset "serialization" begin
-    include("serialization.jl")
-end
